requirements.txt
# MCP (Model Context Protocol) dependencies
mcp>=1.0.0
httpx>=0.25.0
sse-starlette>=1.8.0

# Consul client
python-consul>=1.1.0

# Additional dependencies
pydantic>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0

